requirements.txt
numpy-2.2.6
faiss-gpu-cu12-1.11.0
aiohttp-3.12.6
sentence_transformers-4.1.0
googletrans-4.0.2
fastapi-0.115
uvicorn-0.34.2
any-0.0.2